requirements.txt
fastapi
uvicorn
boto3
pydantic
pillow
requests
python-dotenv
opencv-python
rembg
onnxruntime
